+++ src/style.css
@@ -47,6 +47,31 @@
     text-decoration: none;
     margin: 15px 0;
   }
+
+  /* Define styles for the button-link class */
+.button-link {
+  border-radius: 20px;
+  border: 1px solid #ff4b2b;
+  background-color: #ff4b2b;
+  color: #ffffff;
+  font-size: 10px;
+  font-weight: bold;
+  padding: 6px 25px;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+  transition: transform 80ms ease-in;
+  text-decoration: none; /* Remove underline from the link */
+  display: inline-block; /* Make the link a block-level element */
+  margin-top: 10px; /* Add spacing if needed */
+}
+
+/* Define styles for the button-link when hovered */
+.button-link:hover {
+  transform: scale(1.1); /* Add a hover effect */
+}
+
+
+
   
   button {
     border-radius: 20px;
